Switch from Feign to RestTemplate
It was hard to handle any requests that return a 4xx or 5xx response using
the @FeignClient annotation. RestTemplate can be easily configured to work
with Ribbon using the @LoadBalanced annotation, and provides a lot more
control around handling various responses.

As a side note, in talking with members of the Spring cloud team, the Feign
integration is not fully featured (compared with the native Feign library), so
it seemed that RestTemplate is the more "Spring" like way to go.